+++ src/main/java/nextstep/subway/path/application/DijkstraPathFinder.java
@@ -1,4 +1,4 @@
-package nextstep.subway.path.domain;
+package nextstep.subway.path.application;
 
 import java.util.List;
 
@@ -9,6 +9,7 @@ import org.jgrapht.graph.WeightedMultigraph;
 import org.springframework.stereotype.Component;
 
 import nextstep.subway.line.domain.Sections;
+import nextstep.subway.path.domain.Path;
 import nextstep.subway.station.domain.Station;
 
 @Component
